LibWeb: Split AngleStyleValue out of StyleValue.{h,cpp}
